==> notebooks/cleaning/0_clean_wikipedia.ipynb <==
# Data Cleaning - Wikipedia

# standard library imports
import os

# third party imports
import pandas as pd

# local imports


data_dir = os.path.join(os.path.dirname("__file__"), "..", "..", "data")
raw_data_dir = os.path.join(data_dir, "raw")
clean_data_dir = os.path.join(data_dir, "clean")

## Events

Convert attendance column to nullable integer

wikipedia_events = pd.read_csv(
    os.path.join(raw_data_dir, "Wikipedia", "events.csv"), parse_dates=["date"]
)

# Convert to Int64
wikipedia_events["attendance"] = wikipedia_events["attendance"].astype("Int64")

wikipedia_events.to_csv(
    os.path.join(clean_data_dir, "Wikipedia", "events.csv"), index=False
)

## Venues

No further cleaning needed

wikipedia_venues = pd.read_csv(os.path.join(raw_data_dir, "Wikipedia", "venues.csv"))
wikipedia_venues.to_csv(
    os.path.join(clean_data_dir, "Wikipedia", "venues.csv"), index=False
)
